refactor(trie): move state root database operations into extension trait
