+++ Cargo.toml
@@ -114,6 +114,7 @@ members = [
     "crates/tracing/",
     "crates/transaction-pool/",
     "crates/trie/common",
+    "crates/trie/db",
     "crates/trie/parallel/",
     "crates/trie/trie",
     "examples/beacon-api-sidecar-fetcher/",
@@ -450,9 +451,9 @@ serde_with = "3.3.0"
 humantime = "2.1"
 humantime-serde = "1.1"
 rand = "0.8.5"
-rustc-hash = {  version = "2.0", default-features = false }
+rustc-hash = { version = "2.0", default-features = false }
 schnellru = "0.2"
-strum = {  version = "0.26", default-features = false }
+strum = { version = "0.26", default-features = false }
 rayon = "1.7"
 itertools = "0.13"
 parking_lot = "0.12"
