+++ bin/reth/src/commands/debug_cmd/in_memory_merkle.rs
@@ -27,6 +27,7 @@ use reth_revm::database::StateProviderDatabase;
 use reth_stages::StageId;
 use reth_tasks::TaskExecutor;
 use reth_trie::StateRoot;
+use reth_trie_db::StateRootFromDbTx;
 use std::{path::PathBuf, sync::Arc};
 use tracing::*;
 
@@ -148,8 +149,10 @@ impl Command {
             ExecutionOutcome::new(state, receipts.into(), block.number, vec![requests.into()]);
 
         // Unpacked `BundleState::state_root_slow` function
-        let (in_memory_state_root, in_memory_updates) =
-            execution_outcome.hash_state_slow().state_root_with_updates(provider.tx_ref())?;
+        let (in_memory_state_root, in_memory_updates) = StateRoot::overlay_with_updates(
+            provider.tx_ref(),
+            execution_outcome.hash_state_slow(),
+        )?;
 
         if in_memory_state_root == block.state_root {
             info!(target: "reth::cli", state_root = ?in_memory_state_root, "Computed in-memory state root matches");
