+++ crates/cli/commands/Cargo.toml
@@ -35,6 +35,7 @@ reth-stages.workspace = true
 reth-static-file-types.workspace = true
 reth-static-file.workspace = true
 reth-trie = { workspace = true, features = ["metrics"] }
+reth-trie-db = { workspace = true, features = ["metrics"] }
 
 itertools.workspace = true
 futures.workspace = true
