+++ crates/storage/db-common/Cargo.toml
@@ -16,6 +16,7 @@ reth-db-api.workspace = true
 reth-provider.workspace = true
 reth-config.workspace = true
 reth-trie.workspace = true
+reth-trie-db.workspace = true
 reth-etl.workspace = true
 reth-codecs.workspace = true
 reth-stages-types.workspace = true
