+++ crates/storage/provider/src/providers/state/historical.rs
@@ -15,7 +15,8 @@ use reth_primitives::{
 };
 use reth_storage_api::StateProofProvider;
 use reth_storage_errors::provider::ProviderResult;
-use reth_trie::{updates::TrieUpdates, AccountProof, HashedPostState};
+use reth_trie::{updates::TrieUpdates, AccountProof, HashedPostState, StateRoot};
+use reth_trie_db::StateRootFromDbTx;
 use std::fmt::Debug;
 
 /// State provider for a given block number which takes a tx reference.
@@ -259,7 +260,7 @@ impl<'b, TX: DbTx> StateRootProvider for HistoricalStateProviderRef<'b, TX> {
     fn hashed_state_root(&self, hashed_state: &HashedPostState) -> ProviderResult<B256> {
         let mut revert_state = self.revert_state()?;
         revert_state.extend(hashed_state.clone());
-        revert_state.state_root(self.tx).map_err(|err| ProviderError::Database(err.into()))
+        StateRoot::overlay(self.tx, revert_state).map_err(|err| ProviderError::Database(err.into()))
     }
 
     fn hashed_state_root_with_updates(
@@ -268,8 +269,7 @@ impl<'b, TX: DbTx> StateRootProvider for HistoricalStateProviderRef<'b, TX> {
     ) -> ProviderResult<(B256, TrieUpdates)> {
         let mut revert_state = self.revert_state()?;
         revert_state.extend(hashed_state.clone());
-        revert_state
-            .state_root_with_updates(self.tx)
+        StateRoot::overlay_with_updates(self.tx, revert_state)
             .map_err(|err| ProviderError::Database(err.into()))
     }
 }
