+++ crates/storage/provider/src/providers/state/latest.rs
@@ -12,7 +12,8 @@ use reth_primitives::{
 };
 use reth_storage_api::StateProofProvider;
 use reth_storage_errors::provider::{ProviderError, ProviderResult};
-use reth_trie::{updates::TrieUpdates, AccountProof, HashedPostState};
+use reth_trie::{updates::TrieUpdates, AccountProof, HashedPostState, StateRoot};
+use reth_trie_db::StateRootFromDbTx;
 
 /// State provider over latest state that takes tx reference.
 #[derive(Debug)]
@@ -75,15 +76,15 @@ impl<'b, TX: DbTx> BlockHashReader for LatestStateProviderRef<'b, TX> {
 
 impl<'b, TX: DbTx> StateRootProvider for LatestStateProviderRef<'b, TX> {
     fn hashed_state_root(&self, hashed_state: &HashedPostState) -> ProviderResult<B256> {
-        hashed_state.state_root(self.tx).map_err(|err| ProviderError::Database(err.into()))
+        StateRoot::overlay(self.tx, hashed_state.clone())
+            .map_err(|err| ProviderError::Database(err.into()))
     }
 
     fn hashed_state_root_with_updates(
         &self,
         hashed_state: &HashedPostState,
     ) -> ProviderResult<(B256, TrieUpdates)> {
-        hashed_state
-            .state_root_with_updates(self.tx)
+        StateRoot::overlay_with_updates(self.tx, hashed_state.clone())
             .map_err(|err| ProviderError::Database(err.into()))
     }
 }
