+++ crates/trie/db/src/lib.rs
@@ -1,1 +1,4 @@
+//! An integration of [`reth-trie`] with [`reth-db`].
 
+mod state;
+pub use state::StateRootFromDbTx;
